+++ apis/farms/src/kv.ts
@@ -5,9 +5,22 @@ const KV_PREFIX = {
   lpApr: 'lpApr:',
   farmByPid: 'farmByPid:',
   farmList: 'farmList:',
+  farmV3Liquidity: 'farmV3Liquidity:',
 }
 
 export type FarmResult = Array<FarmWithPrices & { cakeApr?: string; lpApr?: number }>
+
+export type FarmV3LiquidityResult = {
+  tvl: {
+    token0: string
+    token1: string
+  }
+  formatted: {
+    token0: string
+    token1: string
+  }
+  updatedAt: string
+}
 
 export type SavedFarmResult = {
   updatedAt: string
@@ -22,6 +35,7 @@ const createKvKey = {
   apr: (chainId: number | string) => `${KV_PREFIX.lpApr}${chainId}`,
   farm: (chainId: number | string, pid: number | string) => `${KV_PREFIX.farmByPid}${chainId}-${pid}`,
   farms: (chainId: number | string) => `${KV_PREFIX.farmList}${chainId}`,
+  farmsV3Liquidity: (chainId: number | string, address: string) => `${KV_PREFIX.farmV3Liquidity}${chainId}-${address}`,
 }
 
 export class FarmKV {
@@ -48,6 +62,14 @@ export class FarmKV {
       return FARMS.put(createKvKey.apr(chainId), stringifyAprMap)
     }
   }
-}
 
-export const farmKV = new FarmKV()
+  static async getV3Liquidity(chainId: number | string, address: string) {
+    return FARMS.get<FarmV3LiquidityResult>(createKvKey.farmsV3Liquidity(chainId, address), {
+      type: 'json',
+    })
+  }
+
+  static async saveV3Liquidity(chainId: number | string, address: string, result: FarmV3LiquidityResult) {
+    return FARMS.put(createKvKey.farmsV3Liquidity(chainId, address), JSON.stringify(result))
+  }
+}