+++ apps/web/src/utils/shouldShowSwapWarning.ts
@@ -1,9 +1,13 @@
 import { Token } from '@pancakeswap/sdk'
 import SwapWarningTokens from 'config/constants/swapWarningTokens'
 
-const shouldShowSwapWarning = (chainId: number, swapCurrency: Token) => {
-  const swapWarningTokens = Object.values(SwapWarningTokens[chainId])
-  return swapWarningTokens.some((warningToken) => warningToken.equals(swapCurrency))
+const shouldShowSwapWarning = (chainId: number, swapCurrency: Token): boolean => {
+  if (SwapWarningTokens[chainId]) {
+    const swapWarningTokens = Object.values(SwapWarningTokens[chainId])
+    return swapWarningTokens.some((warningToken) => warningToken.equals(swapCurrency))
+  }
+
+  return false
 }
 
 export default shouldShowSwapWarning