+++ apps/web/src/views/Home/components/Banners/hooks/useMultipleBannerConfig.tsx
@@ -1,19 +1,16 @@
 import shuffle from 'lodash/shuffle'
 import { ReactElement, useMemo } from 'react'
-import EthBanner from '../EthBanner'
 import CompetitionBanner from '../CompetitionBanner'
+import { GalxeTraverseBanner } from '../GalxeTraverseBanner'
 import IFOBanner from '../IFOBanner'
-import V3LaunchBanner from '../V3LaunchBanner'
-import PerpetualBanner from '../PerpetualBanner'
 import LiquidStakingBanner from '../LiquidStakingBanner'
-import PancakeProtectorBanner from '../PancakeProtectorBanner'
-import FarmV3MigrationBanner from '../FarmV3MigrationBanner'
+import PerpetualBanner from '../PerpetualBanner'
+import { PolygonZkEvmBanner } from '../PolygonZkEvmBanner'
+import TradingRewardBanner from '../TradingRewardBanner'
+import ArbitrumOneBanner from '../ArbitrumOneBanner'
+import { ZksyncBanner } from '../ZksyncBanner'
 import useIsRenderCompetitionBanner from './useIsRenderCompetitionBanner'
 import useIsRenderIfoBanner from './useIsRenderIFOBanner'
-import TradingRewardBanner from '../TradingRewardBanner'
-import { GalxeTraverseBanner } from '../GalxeTraverseBanner'
-import { PolygonZkEvmBanner } from '../PolygonZkEvmBanner'
-import { ZksyncBanner } from '../ZksyncBanner'
 
 interface IBannerConfig {
   shouldRender: boolean
@@ -39,15 +36,12 @@ export const useMultipleBannerConfig = () => {
 
   return useMemo(() => {
     const NO_SHUFFLE_BANNERS: IBannerConfig[] = [
+      { shouldRender: true, banner: <ArbitrumOneBanner /> },
       { shouldRender: true, banner: <ZksyncBanner /> },
       { shouldRender: true, banner: <PolygonZkEvmBanner /> },
       { shouldRender: true, banner: <GalxeTraverseBanner /> },
-      { shouldRender: true, banner: <PancakeProtectorBanner /> },
       { shouldRender: true, banner: <TradingRewardBanner /> },
       { shouldRender: true, banner: <LiquidStakingBanner /> },
-      { shouldRender: true, banner: <V3LaunchBanner /> },
-      { shouldRender: true, banner: <FarmV3MigrationBanner /> },
-      { shouldRender: true, banner: <EthBanner /> },
       {
         shouldRender: isRenderIFOBanner,
         banner: <IFOBanner />,