+++ packages/smart-router/evm/v3-router/index.ts
@@ -1,2 +1,3 @@
 export * as SmartRouter from './smartRouter'
 export { SwapRouter } from './utils/swapRouter'
+export { Transformer } from './utils'