+++ packages/pools/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pancakeswap/pools
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [e0a681bc6]
+  - @pancakeswap/tokens@0.2.2
+  - @pancakeswap/utils@3.0.2
+  - @pancakeswap/token-lists@0.0.8
+
 ## 4.0.1
 
 ### Patch Changes